+++ polardb-20170801/alibabacloud_polardb20170801/models/_delete_dbendpoint_address_request.py
@@ -24,8 +24,6 @@ class DeleteDBEndpointAddressRequest(DaraModel):
         # The ID of the endpoint.
         # 
         # >  You can call the [DescribeDBClusterEndpoints](https://help.aliyun.com/document_detail/98205.html) operation to query the endpoints of a specified PolarDB cluster.
-        # 
-        # This parameter is required.
         self.dbendpoint_id = dbendpoint_id
         # The network type of the endpoint. Set the value to **Public** (public network).
         #