+++ polardb-20170801/alibabacloud_polardb20170801/models/_modify_dbendpoint_address_request.py
@@ -34,8 +34,6 @@ class ModifyDBEndpointAddressRequest(DaraModel):
         # The ID of the endpoint.
         # 
         # > You can call the [DescribeDBClusterEndpoints](https://help.aliyun.com/document_detail/98205.html) operation to query endpoint IDs.
-        # 
-        # This parameter is required.
         self.dbendpoint_id = dbendpoint_id
         # The network type of the endpoint. Valid values:
         #