+++ polardb-20170801/ChangeLog.md
@@ -1,3 +1,11 @@
+2026-04-08 Version: 6.9.5
+- Update API DescribeDBClusterEndpoints: add response parameters Body.Items.$.ConnectionString.
+- Update API DescribeDBClusterEndpoints: add response parameters Body.Items.$.NetType.
+- Update API DescribeDBClusterEndpoints: add response parameters Body.Items.$.Port.
+- Update API DescribeDBClusterEndpoints: add response parameters Body.Items.$.Protocol.
+- Update API DescribeDBClusterEndpoints: add response parameters Body.Items.$.VPCId.
+
+
 2026-03-30 Version: 6.9.3
 - Update API CreateApplication: add request parameters AuthProvider.
 - Update API CreateApplication: add request parameters AuthProviderConfig.